+++ src/main/java/frc/robot/BuildConstants.java
@@ -7,12 +7,12 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "Reefscape2025";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 67;
-  public static final String GIT_SHA = "0f5d31202e6976bcd143725b5f57e6810caabbff";
-  public static final String GIT_DATE = "2025-02-22 10:33:50 CST";
+  public static final int GIT_REVISION = 68;
+  public static final String GIT_SHA = "a81fdc88f5d6faf56682b4a2c40e386ed3d419aa";
+  public static final String GIT_DATE = "2025-02-22 10:53:24 CST";
   public static final String GIT_BRANCH = "feature/elevator";
-  public static final String BUILD_DATE = "2025-02-22 10:52:24 CST";
-  public static final long BUILD_UNIX_TIME = 1740243144039L;
+  public static final String BUILD_DATE = "2025-02-22 11:38:58 CST";
+  public static final long BUILD_UNIX_TIME = 1740245938034L;
   public static final int DIRTY = 1;
 
   private BuildConstants(){}
