+++ src/main/java/frc/robot/Robot.java
@@ -11,7 +11,8 @@ import org.littletonrobotics.junction.networktables.NT4Publisher;
 import org.littletonrobotics.junction.wpilog.WPILOGReader;
 import org.littletonrobotics.junction.wpilog.WPILOGWriter;
 
-
+import edu.wpi.first.wpilibj.PowerDistribution;
+import edu.wpi.first.wpilibj.PowerDistribution.ModuleType;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 
@@ -32,7 +33,7 @@ public class Robot extends LoggedRobot {
 
       Logger.addDataReceiver(new WPILOGWriter());
       Logger.addDataReceiver(new NT4Publisher());
-      //new PowerDistribution(1, ModuleType.kRev);  //TODO commented out 02-22-2025-REB
+      new PowerDistribution(1, ModuleType.kRev);
     } else {
 
       setUseTiming(false);
