--- src/main/java/frc/robot/subsystems/ElevatorConstants.java
@@ -1,14 +1,0 @@
-package frc.robot.subsystems;
-
-public class ElevatorConstants {
-    
-    public static final int leaderCAN = 38; // TODO Elevator Leader CAN ID
-    public static final int followerCAN = 39; // TODO Elevator Follower CAN ID
-    public static final int currentLimit = 40; // TODO Elevator Current Limit
-    public static final boolean leaderReversed = false;
-    public static final boolean followerReversed = false;
-    public static final int elevatorLaserCAN = 5; //TODO
-
-
-
-}
