+++ src/main/java/frc/robot/subsystems/elevator/ElevatorConstants.java
@@ -1,0 +1,11 @@
+package frc.robot.subsystems.elevator;
+
+public class ElevatorConstants {
+    
+    public static final int leaderCAN = 38; // TODO Elevator Leader CAN ID
+    public static final int followerCAN = 39; // TODO Elevator Follower CAN ID
+    public static final int currentLimit = 40; // TODO Elevator Current Limit
+    public static final boolean leaderReversed = false; // TODO Elevator Leader Reversed
+    public static final boolean followerReversed = false; // TODO Elevator Follower Reversed
+    public static final int laserCAN = 5; //TODO Elevator LaserCAN CAN ID
+}
