+++ src/main/java/frc/robot/subsystems/end_effector/EndEffector.java
@@ -1,4 +1,4 @@
-package frc.robot.end_effector;
+package frc.robot.subsystems.end_effector;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
