+++ src/main/java/frc/robot/subsystems/end_effector/EndEffectorConstants.java
@@ -1,4 +1,4 @@
-package frc.robot.end_effector;
+package frc.robot.subsystems.end_effector;
 
 public class EndEffectorConstants {
